Make datetime.tzinfo non-optional for TZAware (#233)
This PR makes it so that accessing `.tzinfo`, on a `datetime.datetime`
object that has been narrowed to `TZAware`, has the type
`datetime.tzinfo` instead of `datetime.tzinfo | None`. This is safe
because this check is already part of the predicate, ie by definition of
passing `isinstance(_, TZAware)`.

It also addresses #84 for `TZAware` and `TZNaive`, by checking that
accessing an attribute from `datetime.datetime` does not raise a static
type checking error.